Do not use Binding DTO compat methods
This removes the use of compat methods, preferring explicit
instantiation of the correct type.

+++ opendaylight/md-sal/samples/toaster-provider/src/main/java/org/opendaylight/controller/sample/toaster/provider/OpendaylightToaster.java
@@ -66,6 +66,7 @@ import org.opendaylight.yangtools.yang.common.RpcError;
 import org.opendaylight.yangtools.yang.common.RpcError.ErrorType;
 import org.opendaylight.yangtools.yang.common.RpcResult;
 import org.opendaylight.yangtools.yang.common.RpcResultBuilder;
+import org.opendaylight.yangtools.yang.common.Uint16;
 import org.opendaylight.yangtools.yang.common.Uint32;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -101,7 +102,7 @@ public class OpendaylightToaster extends AbstractMXBean
 
     public OpendaylightToaster() {
         this(new ToasterAppConfigBuilder().setManufacturer(TOASTER_MANUFACTURER).setModelNumber(TOASTER_MODEL_NUMBER)
-                .setMaxMakeToastTries(2).build());
+                .setMaxMakeToastTries(Uint16.valueOf(2)).build());
     }
 
     public OpendaylightToaster(final ToasterAppConfig toasterAppConfig) {
